query-service/requirements.txt
# Core dependencies for Query Service
fastapi==0.135.3
uvicorn==0.44.0
python-dotenv==1.2.2
pydantic==2.13.0
sqlalchemy==2.0.49
asyncpg==0.31.0
greenlet==3.4.0

# AI/LangChain dependencies
langchain==1.2.15
langchain-openai==1.1.12
langchain-chroma==1.1.0
langchain-core==1.2.28
langchain-classic==1.0.3
